Refine CardAggregate API
